refactored much of the toolchain

+++ unload_log_tcp_cwnd.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
-BPFTOOL=/home/lakshay21060/bpftool/src/bpftool
+if [ -z "${BPFTOOL}" ]; then
+    echo "error: BPFTOOL is not set"
+    exit 1
+fi
+
 if [ -f "/sys/fs/bpf/log_tcp_cwnd" ]; then
-	sudo ${BPFTOOL} prog detach pinned /sys/fs/bpf/log_tcp_cwnd tracepoint
+	sudo $BPFTOOL prog detach pinned /sys/fs/bpf/log_tcp_cwnd tracepoint
 	rm -f /sys/fs/bpf/log_tcp_cwnd
 fi
